fix Richardson-Lucy args ; fix bug when saving workflow without Data folder

+++ PyFlow/Packages/PyFlowBase/Tools/WorkflowTool.py
@@ -302,10 +302,11 @@ class WorkflowTool(DockTool):
             # Copy parameters.json and output_data_frame.csv files (PARAMETERS_PATH, OUTPUT_DATAFRAME_PATH) in the "Data / nodeName" folders, but not the other data files
             dataFolder = workflowFolder / 'Data'
             dataFolder.mkdir()
-            for folder in sorted(list((path / 'Data').iterdir())):
-                (dataFolder / folder.name).mkdir()
-                for fileName in [PARAMETERS_PATH, OUTPUT_DATAFRAME_PATH]:
-                    shutil.copyfile(folder / fileName, dataFolder / folder.name / fileName)
+            if (path / 'Data').exists():
+                for folder in sorted(list((path / 'Data').iterdir())):
+                    (dataFolder / folder.name).mkdir()
+                    for fileName in [PARAMETERS_PATH, OUTPUT_DATAFRAME_PATH]:
+                        shutil.copyfile(folder / fileName, dataFolder / folder.name / fileName)
             shutil.copyfile(path / WorkflowTool.graphFileName, workflowFolder / WorkflowTool.graphFileName)
             shutil.make_archive(zipFilePath.with_suffix(''), 'zip', tmp)
         return
